remove non-native grpc client
